Add files via upload
i.	Introduction: 
The Abalone dataset is a popular dataset in the field of machine learning, which is commonly used for regression analysis. The dataset is related to the age of the Abalone, which is a type of marine snail. The dataset contains information about the physical measurements of the Abalone, such as the length, diameter, and weight, as well as its age. The aim of this report is to analyze the Abalone dataset using regression analysis and to discuss the results.
ii.	Methodology: 

The dataset used in this analysis contains 4177 observations and 9 variables. The dataset was preprocessed by encoding the gender variable and splitting the dataset into training and testing sets. The training set contained 70% of the data, and the testing set contained 30% of the data. The regression model used in this analysis was the linear regression model. The model was fitted using the training set, and the performance was evaluated using the testing set. The performance of the model was evaluated using four different metrics, namely, the mean squared error (MSE), the coefficient of determination R-squared (R2), the root mean squared error (RMSE), and the mean absolute error (MAE).
Results: The performance of the 
1.	linear regression model was evaluated using the testing set, and the following results were obtained:
•	Mean Squared Error (MSE): 5.585855026952151
•	Coefficient of determination R-squared (R2): 0.4799089551478021
•	Root Mean Squared Error (RMSE): 2.36344135255185
•	Mean Absolute Error (MAE): 2.36344135255185
2.	With Neural Network model was evaluated using the testing set, and the following results were obtained:
•	Mean Squared Error (MSE): 4.532966136932373
•	Coefficient of determination R-squared (R2): 0.5779419097025543
•	Root Mean Squared Error (RMSE): 2.12907635770358
•	Mean Absolute Error (MAE): 1.5163860321044922.

3.	With XG Boost model was evaluated using the testing set, and the following results were obtained:
•	Mean Squared Error (MSE): 5.557944584159364
•	Coefficient of determination R-squared (R2): 0.48250765691939257
•	Root Mean Squared Error (RMSE): 0.48250765691939257
•	Mean Absolute Error (MAE): 1.66271550041472
Discussion: The results obtained from the regression analysis, xg boost and neural network show that the linear regression model is not very accurate in predicting the age of the Abalone based on its physical measurements. This is evident from the relatively high values of the MSE, RMSE, and MAE. The R-squared value of 0.4799 indicates that the model explains only 47.99% of the variation in the response variable if we use normal Linear Regression. The R-squared value of 0.5779 indicates that the model explains only 57.79% of the variation in the response variable if we use. The R-squared value of 0.4825 indicates that the model explains only 48.25% of the variation in the response variable if we use. 
Therefore, it can be concluded that the linear regression model and xg boost is not a suitable model for predicting the age of the Abalone based on its physical measurements however if we create a neural network and trained on it then the performance of the model will better compared to those two.
In conclusion, the Abalone dataset was analyzed using regression analysis, and the linear regression model was used to predict the age of the Abalone based on its physical measurements. The results obtained from the analysis show that the linear regression model is not very accurate in predicting the age of the Abalone and the model performed moderately if we include the use of neural network. Further analysis using different regression models or additional variables may provide better results.